emotion_ml/requirements.txt
concrete-ml>=1.9.0
scikit-learn>=1.3.0
pandas>=2.0
datasets>=2.14
joblib>=1.3
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
